+++ packages/windie-sdk-js/cjs/transport/BackendSocketFactory.js
@@ -3,8 +3,8 @@
  * Provides the backend socket factory module for the TypeScript SDK runtime.
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.createWindieSdkBackendSocket = createWindieSdkBackendSocket;
-function createWindieSdkBackendSocket({ WebSocketImpl, wsUrl, wsOrigin, headers, }) {
+exports.createWindieSdkBackendSocket = exports.createAgentBackendSocket = void 0;
+function createAgentBackendSocket({ WebSocketImpl, wsUrl, wsOrigin, headers, }) {
     if (!WebSocketImpl) {
         throw new Error('Agent SDK backend socket requires WebSocketImpl');
     }
@@ -16,3 +16,5 @@ function createWindieSdkBackendSocket({ WebSocketImpl, wsUrl, wsOrigin, headers,
         headers: headers || {},
     });
 }
+exports.createAgentBackendSocket = createAgentBackendSocket;
+exports.createWindieSdkBackendSocket = createAgentBackendSocket;